feat: click listeners to view holders

+++ app/src/main/java/cherryjam/narfu/arkhdialect/adapter/PhotoAttachmentAdapter.kt
@@ -1,8 +1,8 @@
 package cherryjam.narfu.arkhdialect.adapter
 
+import android.content.Intent
 import android.util.Size
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import android.widget.PopupMenu
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
@@ -13,7 +13,6 @@ import cherryjam.narfu.arkhdialect.databinding.ItemPhotoAttachmentBinding
 
 class PhotoAttachmentAdapter :
     SelectableAdapter<PhotoAttachmentAdapter.PhotoAttachmentViewHolder>() {
-
     var data: List<PhotoAttachment> = emptyList()
         set(newValue) {
             field = newValue
@@ -27,27 +26,37 @@ class PhotoAttachmentAdapter :
 
         init {
             binding.imageView.setOnClickListener {
-                val popup = PopupMenu(it.context, it)
-                popup.inflate(R.menu.options_menu)
+                if (isSelecting)
+                    selectItem(this)
+                else {
+                    val popup = PopupMenu(it.context, it)
+                    popup.inflate(R.menu.options_menu)
 
-                popup.setOnMenuItemClickListener { item ->
-                    when (item.itemId) {
-                        R.id.select -> {
-//                            if (!isSelecting)
-//                                startSelection()
-//                            selectItem(this)
-                            true
+                    popup.setOnMenuItemClickListener { item ->
+                        when (item.itemId) {
+                            R.id.select -> {
+                                if (!isSelecting)
+                                    startSelection()
+                                selectItem(this)
+                                true
+                            }
+                            R.id.delete -> {
+                                Thread {
+                                    AppDatabase.getInstance().photoAttachmentDao().delete(attachment)
+                                }.start()
+                                true//
+                            }
+                            else -> false
                         }
-                        R.id.delete -> {
-                            Thread {
-                                AppDatabase.getInstance().photoAttachmentDao().delete(attachment)
-                            }.start()
-                            true//
-                        }
-                        else -> false
                     }
+                    popup.show()
                 }
-                popup.show()
+            }
+            binding.imageView.setOnLongClickListener {
+                if (!isSelecting)
+                    startSelection()
+                selectItem(this)
+                true
             }
         }
 
@@ -76,6 +85,14 @@ class PhotoAttachmentAdapter :
         override fun onDeselect() {
             binding.imageView.scaleX = 1f
             binding.imageView.scaleY = 1f
+        }
+
+        fun openAttachment() {
+            val intent = Intent().apply {
+                setAction(Intent.ACTION_VIEW)
+                setDataAndType(attachment.uri, context.contentResolver.getType(attachment.uri))
+            }
+            context.startActivity(intent)
         }
     }
